add confirm dialog for delete operations

+++ src/app/admin-detail/admin-detail.component.ts
@@ -1,14 +1,15 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { Workshop } from '../Workshop';
-import { ActivatedRoute } from '../../../node_modules/@angular/router';
-import { WorkshopService } from '../workshop.service';
-import { ParticipantService } from '../participant.service';
-import { Participant } from '../Participant';
+import { Component, OnInit, Input, TemplateRef } from "@angular/core";
+import { Workshop } from "../Workshop";
+import { ActivatedRoute } from "../../../node_modules/@angular/router";
+import { WorkshopService } from "../workshop.service";
+import { ParticipantService } from "../participant.service";
+import { Participant } from "../Participant";
+import { BsModalService, BsModalRef } from "ngx-bootstrap/modal";
 
 @Component({
-  selector: 'app-admin-detail',
-  templateUrl: './admin-detail.component.html',
-  styleUrls: ['./admin-detail.component.scss']
+  selector: "app-admin-detail",
+  templateUrl: "./admin-detail.component.html",
+  styleUrls: ["./admin-detail.component.scss"]
 })
 export class AdminDetailComponent implements OnInit {
   @Input()
@@ -38,8 +39,9 @@ export class AdminDetailComponent implements OnInit {
   }
 
   deleteParticipant(participant: Participant) {
-    this.participants = this.participants.filter(p => p !== participant);
-    this.participantService.deleteParticipant(participant).subscribe();
+    if (confirm("Are you sure?")) {
+      this.participants = this.participants.filter(p => p !== participant);
+      this.participantService.deleteParticipant(participant).subscribe();
+    }
   }
-
 }
